Refactoring integration tests Tests pass now
