+++ JustSaying.IntegrationTests/TestHandlers/OrderDispatcher.cs
@@ -1,0 +1,26 @@
+using System.Threading.Tasks;
+using JustSaying.Messaging.MessageHandling;
+
+namespace JustSaying.IntegrationTests.TestHandlers
+{
+    public class OrderDispatcher : IHandlerAsync<OrderPlaced>
+    {
+        private readonly Future<OrderPlaced> _future;
+
+        public OrderDispatcher(Future<OrderPlaced> future)
+        {
+            _future = future;
+        }
+
+        public async Task<bool> Handle(OrderPlaced message)
+        {
+            await _future.Complete(message);
+            return true;
+        }
+
+        public Future<OrderPlaced> Future
+        {
+            get { return _future; }
+        }
+    }
+}
