+++ JustSaying.IntegrationTests/TestHandlers/OrderPlaced.cs
@@ -1,0 +1,13 @@
+using JustSaying.Models;
+
+namespace JustSaying.IntegrationTests.TestHandlers
+{
+    public class OrderPlaced : Message
+    {
+        public OrderPlaced(string orderId)
+        {
+            OrderId = orderId;
+        }
+        public string OrderId { get; private set; }
+    }
+}
