+++ JustSaying.IntegrationTests/TestHandlers/ThrowingHandler.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using JustSaying.Messaging.MessageHandling;
 using JustSaying.TestingFramework;
 
-namespace JustSaying.IntegrationTests.JustSayingFluently
+namespace JustSaying.IntegrationTests.TestHandlers
 {
     public class ThrowingHandler : IHandlerAsync<GenericMessage>
     {
