+++ JustSaying.IntegrationTests/WhenRegisteringHandlersViaResolver/MultipleHandlerRegistry.cs
@@ -1,4 +1,5 @@
 using JustSaying.IntegrationTests.JustSayingFluently;
+using JustSaying.IntegrationTests.TestHandlers;
 using JustSaying.Messaging.MessageHandling;
 using StructureMap.Configuration.DSL;
