+++ JustSaying.IntegrationTests/WhenRegisteringHandlersViaResolver/WhenRegisteringABlockingHandlerViaContainer.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using JustSaying.IntegrationTests.TestHandlers;
 using JustSaying.Messaging.MessageHandling;
 using NUnit.Framework;
 using Shouldly;
