+++ JustSaying.IntegrationTests/WhenRegisteringHandlersViaResolver/WhenRegisteringASingleHandlerViaContainer.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using JustSaying.IntegrationTests.JustSayingFluently;
+using JustSaying.IntegrationTests.TestHandlers;
 using NUnit.Framework;
 using Shouldly;
 using StructureMap;
